+++ sdks/java/io/google-cloud-platform/src/main/java/org/apache/beam/sdk/io/gcp/datastore/DatastoreV1.java
@@ -66,6 +66,7 @@ import org.apache.beam.sdk.coders.StringUtf8Coder;
 import org.apache.beam.sdk.extensions.gcp.options.GcpOptions;
 import org.apache.beam.sdk.extensions.gcp.util.RetryHttpRequestInitializer;
 import org.apache.beam.sdk.metrics.Counter;
+import org.apache.beam.sdk.metrics.Distribution;
 import org.apache.beam.sdk.metrics.Metrics;
 import org.apache.beam.sdk.options.PipelineOptions;
 import org.apache.beam.sdk.options.ValueProvider;
@@ -166,6 +167,13 @@ import org.slf4j.LoggerFactory;
  * p.run();
  * }</pre>
  *
+ * <p>Write and delete operations will follow a gradual ramp-up by default in order to protect Cloud
+ * Datastore from potential overload. This rate limit follows a heuristic based on the expected
+ * number of workers. To optimize throughput in this initial stage, you can provide a hint to the
+ * relevant {@code PTransform} by calling {@code withHintNumWorkers}, e.g., {@code
+ * DatastoreIO.v1().deleteKey().withHintNumWorkers(numWorkers)}. While not recommended, you can also
+ * turn this off via {@code .withRampupThrottlingDisabled()}.
+ *
  * <p>{@link Entity Entities} in the {@code PCollection} to be written or deleted must have complete
  * {@link Key Keys}. Complete {@code Keys} specify the {@code name} and {@code id} of the {@code
  * Entity}, where incomplete {@code Keys} do not. A {@code namespace} other than {@code projectId}
@@ -209,10 +217,10 @@ public class DatastoreV1 {
    * size; it is adjusted at runtime based on the performance of previous writes (see {@link
    * DatastoreV1.WriteBatcher}).
    *
-   * <p>Testing has found that a batch of 200 entities will generally finish within the timeout even
+   * <p>Testing has found that a batch of 50 entities will generally finish within the timeout even
    * in adverse conditions.
    */
-  @VisibleForTesting static final int DATASTORE_BATCH_UPDATE_ENTITIES_START = 200;
+  @VisibleForTesting static final int DATASTORE_BATCH_UPDATE_ENTITIES_START = 50;
 
   /**
    * When choosing the number of updates in a single RPC, never exceed the maximum allowed by the
@@ -225,7 +233,7 @@ public class DatastoreV1 {
    * number of entities per request may be lower when we flush for the end of a bundle or if we hit
    * {@link DatastoreV1.DATASTORE_BATCH_UPDATE_BYTES_LIMIT}.
    */
-  @VisibleForTesting static final int DATASTORE_BATCH_UPDATE_ENTITIES_MIN = 10;
+  @VisibleForTesting static final int DATASTORE_BATCH_UPDATE_ENTITIES_MIN = 5;
 
   /**
    * Cloud Datastore has a limit of 10MB per RPC, so we also flush if the total size of mutations
@@ -235,7 +243,14 @@ public class DatastoreV1 {
   @VisibleForTesting static final int DATASTORE_BATCH_UPDATE_BYTES_LIMIT = 9_000_000;
 
   /**
-   * Non-retryable errors. See https://cloud.google.com/datastore/docs/concepts/errors#Error_Codes .
+   * Default hint for the expected number of workers in the ramp-up throttling step. This is used to
+   * approximate a global rate limit, so a mismatch can yield slightly elevated throttling.
+   * Reconfigure on the {@link Mutate} object to align with the worker count and improve throughput.
+   */
+  private static final int DEFAULT_HINT_NUM_WORKERS = 500;
+
+  /**
+   * Non-retryable errors. See https://cloud.google.com/datastore/docs/concepts/errors#Error_Codes.
    */
   private static final Set<Code> NON_RETRYABLE_ERRORS =
       ImmutableSet.of(
@@ -970,7 +985,7 @@ public class DatastoreV1 {
    * using {@link DatastoreV1.Write#withProjectId}.
    */
   public Write write() {
-    return new Write(null, null);
+    return new Write(null, null, true, DEFAULT_HINT_NUM_WORKERS);
   }
 
   /**
@@ -978,7 +993,7 @@ public class DatastoreV1 {
    * using {@link DeleteEntity#withProjectId}.
    */
   public DeleteEntity deleteEntity() {
-    return new DeleteEntity(null, null);
+    return new DeleteEntity(null, null, true, DEFAULT_HINT_NUM_WORKERS);
   }
 
   /**
@@ -986,7 +1001,7 @@ public class DatastoreV1 {
    * {@link DeleteKey#withProjectId}.
    */
   public DeleteKey deleteKey() {
-    return new DeleteKey(null, null);
+    return new DeleteKey(null, null, true, DEFAULT_HINT_NUM_WORKERS);
   }
 
   /**
@@ -995,12 +1010,17 @@ public class DatastoreV1 {
    * @see DatastoreIO
    */
   public static class Write extends Mutate<Entity> {
+
     /**
      * Note that {@code projectId} is only {@code @Nullable} as a matter of build order, but if it
      * is {@code null} at instantiation time, an error will be thrown.
      */
-    Write(@Nullable ValueProvider<String> projectId, @Nullable String localhost) {
-      super(projectId, localhost, new UpsertFn());
+    Write(
+        @Nullable ValueProvider<String> projectId,
+        @Nullable String localhost,
+        boolean throttleRampup,
+        int hintNumWorkers) {
+      super(projectId, localhost, new UpsertFn(), throttleRampup, hintNumWorkers);
     }
 
     /** Returns a new {@link Write} that writes to the Cloud Datastore for the specified project. */
@@ -1012,7 +1032,7 @@ public class DatastoreV1 {
     /** Same as {@link Write#withProjectId(String)} but with a {@link ValueProvider}. */
     public Write withProjectId(ValueProvider<String> projectId) {
       checkArgument(projectId != null, "projectId can not be null");
-      return new Write(projectId, localhost);
+      return new Write(projectId, localhost, throttleRampup, hintNumWorkers);
     }
 
     /**
@@ -1021,7 +1041,21 @@ public class DatastoreV1 {
      */
     public Write withLocalhost(String localhost) {
       checkArgument(localhost != null, "localhost can not be null");
-      return new Write(projectId, localhost);
+      return new Write(projectId, localhost, throttleRampup, hintNumWorkers);
+    }
+
+    /** Returns a new {@link Write} that does not throttle during ramp-up. */
+    public Write withRampupThrottlingDisabled() {
+      return new Write(projectId, localhost, false, hintNumWorkers);
+    }
+
+    /**
+     * Returns a new {@link Write} with a different worker count hint for ramp-up throttling. Value
+     * is ignored if ramp-up throttling is disabled.
+     */
+    public Write withHintNumWorkers(int hintNumWorkers) {
+      checkArgument(hintNumWorkers > 0, "hintNumWorkers must be positive");
+      return new Write(projectId, localhost, throttleRampup, hintNumWorkers);
     }
   }
 
@@ -1031,12 +1065,17 @@ public class DatastoreV1 {
    * @see DatastoreIO
    */
   public static class DeleteEntity extends Mutate<Entity> {
+
     /**
      * Note that {@code projectId} is only {@code @Nullable} as a matter of build order, but if it
      * is {@code null} at instantiation time, an error will be thrown.
      */
-    DeleteEntity(@Nullable ValueProvider<String> projectId, @Nullable String localhost) {
-      super(projectId, localhost, new DeleteEntityFn());
+    DeleteEntity(
+        @Nullable ValueProvider<String> projectId,
+        @Nullable String localhost,
+        boolean throttleRampup,
+        int hintNumWorkers) {
+      super(projectId, localhost, new DeleteEntityFn(), throttleRampup, hintNumWorkers);
     }
 
     /**
@@ -1051,7 +1090,7 @@ public class DatastoreV1 {
     /** Same as {@link DeleteEntity#withProjectId(String)} but with a {@link ValueProvider}. */
     public DeleteEntity withProjectId(ValueProvider<String> projectId) {
       checkArgument(projectId != null, "projectId can not be null");
-      return new DeleteEntity(projectId, localhost);
+      return new DeleteEntity(projectId, localhost, throttleRampup, hintNumWorkers);
     }
 
     /**
@@ -1060,7 +1099,21 @@ public class DatastoreV1 {
      */
     public DeleteEntity withLocalhost(String localhost) {
       checkArgument(localhost != null, "localhost can not be null");
-      return new DeleteEntity(projectId, localhost);
+      return new DeleteEntity(projectId, localhost, throttleRampup, hintNumWorkers);
+    }
+
+    /** Returns a new {@link DeleteEntity} that does not throttle during ramp-up. */
+    public DeleteEntity withRampupThrottlingDisabled() {
+      return new DeleteEntity(projectId, localhost, false, hintNumWorkers);
+    }
+
+    /**
+     * Returns a new {@link DeleteEntity} with a different worker count hint for ramp-up throttling.
+     * Value is ignored if ramp-up throttling is disabled.
+     */
+    public DeleteEntity withHintNumWorkers(int hintNumWorkers) {
+      checkArgument(hintNumWorkers > 0, "hintNumWorkers must be positive");
+      return new DeleteEntity(projectId, localhost, throttleRampup, hintNumWorkers);
     }
   }
 
@@ -1071,12 +1124,17 @@ public class DatastoreV1 {
    * @see DatastoreIO
    */
   public static class DeleteKey extends Mutate<Key> {
+
     /**
      * Note that {@code projectId} is only {@code @Nullable} as a matter of build order, but if it
      * is {@code null} at instantiation time, an error will be thrown.
      */
-    DeleteKey(@Nullable ValueProvider<String> projectId, @Nullable String localhost) {
-      super(projectId, localhost, new DeleteKeyFn());
+    DeleteKey(
+        @Nullable ValueProvider<String> projectId,
+        @Nullable String localhost,
+        boolean throttleRampup,
+        int hintNumWorkers) {
+      super(projectId, localhost, new DeleteKeyFn(), throttleRampup, hintNumWorkers);
     }
 
     /**
@@ -1094,13 +1152,27 @@ public class DatastoreV1 {
      */
     public DeleteKey withLocalhost(String localhost) {
       checkArgument(localhost != null, "localhost can not be null");
-      return new DeleteKey(projectId, localhost);
+      return new DeleteKey(projectId, localhost, throttleRampup, hintNumWorkers);
     }
 
     /** Same as {@link DeleteKey#withProjectId(String)} but with a {@link ValueProvider}. */
     public DeleteKey withProjectId(ValueProvider<String> projectId) {
       checkArgument(projectId != null, "projectId can not be null");
-      return new DeleteKey(projectId, localhost);
+      return new DeleteKey(projectId, localhost, throttleRampup, hintNumWorkers);
+    }
+
+    /** Returns a new {@link DeleteKey} that does not throttle during ramp-up. */
+    public DeleteKey withRampupThrottlingDisabled() {
+      return new DeleteKey(projectId, localhost, false, hintNumWorkers);
+    }
+
+    /**
+     * Returns a new {@link DeleteKey} with a different worker count hint for ramp-up throttling.
+     * Value is ignored if ramp-up throttling is disabled.
+     */
+    public DeleteKey withHintNumWorkers(int hintNumWorkers) {
+      checkArgument(hintNumWorkers > 0, "hintNumWorkers must be positive");
+      return new DeleteKey(projectId, localhost, throttleRampup, hintNumWorkers);
     }
   }
 
@@ -1113,10 +1185,15 @@ public class DatastoreV1 {
    * provided, as the commits are retried when failures occur.
    */
   private abstract static class Mutate<T> extends PTransform<PCollection<T>, PDone> {
+
     protected ValueProvider<String> projectId;
     protected @Nullable String localhost;
+    protected boolean throttleRampup;
+    protected int hintNumWorkers;
     /** A function that transforms each {@code T} into a mutation. */
     private final SimpleFunction<T, Mutation> mutationFn;
+
+    private final RampupThrottlingFn<Mutation> rampupThrottlingFn;
 
     /**
      * Note that {@code projectId} is only {@code @Nullable} as a matter of build order, but if it
@@ -1125,10 +1202,15 @@ public class DatastoreV1 {
     Mutate(
         @Nullable ValueProvider<String> projectId,
         @Nullable String localhost,
-        SimpleFunction<T, Mutation> mutationFn) {
+        SimpleFunction<T, Mutation> mutationFn,
+        boolean throttleRampup,
+        int hintNumWorkers) {
       this.projectId = projectId;
       this.localhost = localhost;
+      this.throttleRampup = throttleRampup;
+      this.hintNumWorkers = hintNumWorkers;
       this.mutationFn = checkNotNull(mutationFn);
+      this.rampupThrottlingFn = new RampupThrottlingFn<>(hintNumWorkers);
     }
 
     @Override
@@ -1139,10 +1221,15 @@ public class DatastoreV1 {
       }
       checkArgument(mutationFn != null, "mutationFn can not be null");
 
-      input
-          .apply("Convert to Mutation", MapElements.via(mutationFn))
-          .apply(
-              "Write Mutation to Datastore", ParDo.of(new DatastoreWriterFn(projectId, localhost)));
+      PCollection<Mutation> intermediateOutput =
+          input.apply("Convert to Mutation", MapElements.via(mutationFn));
+      if (throttleRampup) {
+        intermediateOutput =
+            intermediateOutput.apply(
+                "Enforce ramp-up through throttling", ParDo.of(rampupThrottlingFn));
+      }
+      intermediateOutput.apply(
+          "Write Mutation to Datastore", ParDo.of(new DatastoreWriterFn(projectId, localhost)));
 
       return PDone.in(input.getPipeline());
     }
@@ -1161,6 +1248,9 @@ public class DatastoreV1 {
       builder
           .addIfNotNull(DisplayData.item("projectId", projectId).withLabel("Output Project"))
           .include("mutationFn", mutationFn);
+      if (throttleRampup) {
+        builder.include("rampupThrottlingFn", rampupThrottlingFn);
+      }
     }
 
     public String getProjectId() {
@@ -1171,6 +1261,7 @@ public class DatastoreV1 {
   /** Determines batch sizes for commit RPCs. */
   @VisibleForTesting
   interface WriteBatcher {
+
     /** Call before using this WriteBatcher. */
     void start();
 
@@ -1195,8 +1286,9 @@ public class DatastoreV1 {
    */
   @VisibleForTesting
   static class WriteBatcherImpl implements WriteBatcher, Serializable {
+
     /** Target time per RPC for writes. */
-    static final int DATASTORE_BATCH_TARGET_LATENCY_MS = 5000;
+    static final int DATASTORE_BATCH_TARGET_LATENCY_MS = 6000;
 
     @Override
     public void start() {
@@ -1217,12 +1309,11 @@ public class DatastoreV1 {
         return DATASTORE_BATCH_UPDATE_ENTITIES_START;
       }
       long recentMeanLatency = Math.max(meanLatencyPerEntityMs.get(timeSinceEpochMillis), 1);
+      long targetBatchSize = DATASTORE_BATCH_TARGET_LATENCY_MS / recentMeanLatency;
       return (int)
           Math.max(
               DATASTORE_BATCH_UPDATE_ENTITIES_MIN,
-              Math.min(
-                  DATASTORE_BATCH_UPDATE_ENTITIES_LIMIT,
-                  DATASTORE_BATCH_TARGET_LATENCY_MS / recentMeanLatency));
+              Math.min(DATASTORE_BATCH_UPDATE_ENTITIES_LIMIT, targetBatchSize));
     }
 
     private transient MovingAverage meanLatencyPerEntityMs;
@@ -1243,6 +1334,7 @@ public class DatastoreV1 {
    */
   @VisibleForTesting
   static class DatastoreWriterFn extends DoFn<Mutation, Void> {
+
     private static final Logger LOG = LoggerFactory.getLogger(DatastoreWriterFn.class);
     private final ValueProvider<String> projectId;
     private final @Nullable String localhost;
@@ -1252,13 +1344,17 @@ public class DatastoreV1 {
     private final List<Mutation> mutations = new ArrayList<>();
     private int mutationsSize = 0; // Accumulated size of protos in mutations.
     private WriteBatcher writeBatcher;
-    private transient AdaptiveThrottler throttler;
+    private transient AdaptiveThrottler adaptiveThrottler;
     private final Counter throttlingMsecs =
         Metrics.counter(DatastoreWriterFn.class, "throttling-msecs");
     private final Counter rpcErrors =
         Metrics.counter(DatastoreWriterFn.class, "datastoreRpcErrors");
     private final Counter rpcSuccesses =
         Metrics.counter(DatastoreWriterFn.class, "datastoreRpcSuccesses");
+    private final Counter entitiesMutated =
+        Metrics.counter(DatastoreWriterFn.class, "datastoreEntitiesMutated");
+    private final Distribution latencyMsPerMutation =
+        Metrics.distribution(DatastoreWriterFn.class, "datastoreLatencyMsPerMutation");
 
     private static final int MAX_RETRIES = 5;
     private static final FluentBackoff BUNDLE_WRITE_BACKOFF =
@@ -1294,9 +1390,9 @@ public class DatastoreV1 {
     public void startBundle(StartBundleContext c) {
       datastore = datastoreFactory.getDatastore(c.getPipelineOptions(), projectId.get(), localhost);
       writeBatcher.start();
-      if (throttler == null) {
+      if (adaptiveThrottler == null) {
         // Initialize throttler at first use, because it is not serializable.
-        throttler = new AdaptiveThrottler(120000, 10000, 1.25);
+        adaptiveThrottler = new AdaptiveThrottler(120000, 10000, 1.25);
       }
     }
 
@@ -1344,7 +1440,7 @@ public class DatastoreV1 {
         commitRequest.setMode(CommitRequest.Mode.NON_TRANSACTIONAL);
         long startTime = System.currentTimeMillis(), endTime;
 
-        if (throttler.throttleRequest(startTime)) {
+        if (adaptiveThrottler.throttleRequest(startTime)) {
           LOG.info("Delaying request due to previous failures");
           throttlingMsecs.inc(WriteBatcherImpl.DATASTORE_BATCH_TARGET_LATENCY_MS);
           sleeper.sleep(WriteBatcherImpl.DATASTORE_BATCH_TARGET_LATENCY_MS);
@@ -1356,8 +1452,10 @@ public class DatastoreV1 {
           endTime = System.currentTimeMillis();
 
           writeBatcher.addRequestLatency(endTime, endTime - startTime, mutations.size());
-          throttler.successfulRequest(startTime);
+          adaptiveThrottler.successfulRequest(startTime);
+          latencyMsPerMutation.update((endTime - startTime) / mutations.size());
           rpcSuccesses.inc();
+          entitiesMutated.inc(mutations.size());
 
           // Break if the commit threw no exception.
           break;
@@ -1368,6 +1466,7 @@ public class DatastoreV1 {
              * consideration, though. */
             endTime = System.currentTimeMillis();
             writeBatcher.addRequestLatency(endTime, endTime - startTime, mutations.size());
+            latencyMsPerMutation.update((endTime - startTime) / mutations.size());
           }
           // Only log the code and message for potentially-transient errors. The entire exception
           // will be propagated upon the last retry.
@@ -1497,13 +1596,21 @@ public class DatastoreV1 {
     public Datastore getDatastore(
         PipelineOptions pipelineOptions, String projectId, @Nullable String localhost) {
       Credentials credential = pipelineOptions.as(GcpOptions.class).getGcpCredential();
+
+      // Add Beam version to user agent header.
+      HttpRequestInitializer userAgentInitializer =
+          request -> request.getHeaders().setUserAgent(pipelineOptions.getUserAgent());
       HttpRequestInitializer initializer;
       if (credential != null) {
         initializer =
             new ChainingHttpRequestInitializer(
-                new HttpCredentialsAdapter(credential), new RetryHttpRequestInitializer());
+                new HttpCredentialsAdapter(credential),
+                new RetryHttpRequestInitializer(),
+                userAgentInitializer);
       } else {
-        initializer = new RetryHttpRequestInitializer();
+        initializer =
+            new ChainingHttpRequestInitializer(
+                new RetryHttpRequestInitializer(), userAgentInitializer);
       }
 
       DatastoreOptions.Builder builder =